dismech-workspace/benchmark-rag-vs-structured/pyproject.toml
[project]
name = "dismech-benchmark"
version = "0.1.0"
description = "RAG vs. Structured Knowledge benchmark for the DisMech disease mechanism KB"
requires-python = ">=3.11,<3.14"
dependencies = [
    "anthropic>=0.49.0",
    "voyageai>=0.3.0",
    "qdrant-client>=1.9.0",
    "pyyaml>=6.0.0",
    "matplotlib>=3.8.0",
    "requests>=2.31.0",
    "typedb-driver>=3.8.0",
    "tqdm>=4.0.0",
]

[tool.uv]
dev-dependencies = []
